+++ chpl/chpl-auth/src/main/java/gov/healthit/chpl/auth/Util.java
@@ -66,20 +66,7 @@ public class Util {
 			return false;
 		}
 		for (GrantedPermission perm : user.getPermissions()) {
-			if (perm.getAuthority().equals(Authority.ROLE_ACB_ADMIN)) {
-				return true;
-			}
-		}
-		return false;
-	}
-
-	public static boolean isUserRoleAcbStaff() {
-		User user = getCurrentUser();
-		if (user == null) {
-			return false;
-		}
-		for (GrantedPermission perm : user.getPermissions()) {
-			if (perm.getAuthority().equals(Authority.ROLE_ACB_STAFF)) {
+			if (perm.getAuthority().equals(Authority.ROLE_ACB)) {
 				return true;
 			}
 		}