+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/manager/impl/CertifiedProductManagerImpl.java
@@ -70,7 +70,6 @@ import gov.healthit.chpl.domain.IcsFamilyTreeNode;
 import gov.healthit.chpl.domain.InheritedCertificationStatus;
 import gov.healthit.chpl.domain.ListingUpdateRequest;
 import gov.healthit.chpl.domain.MeaningfulUseUser;
-import gov.healthit.chpl.domain.concept.ActivityConcept;
 import gov.healthit.chpl.dto.AccessibilityStandardDTO;
 import gov.healthit.chpl.dto.AddressDTO;
 import gov.healthit.chpl.dto.CQMCriterionDTO;
@@ -378,7 +377,7 @@ public class CertifiedProductManagerImpl implements CertifiedProductManager {
     }
 
     @Override
-    @PreAuthorize("(hasRole('ROLE_ACB_STAFF') or hasRole('ROLE_ACB_ADMIN')) "
+    @PreAuthorize("hasRole('ROLE_ACB') "
             + "and hasPermission(#acbId, 'gov.healthit.chpl.dto.CertificationBodyDTO', admin)")
     @Transactional(readOnly = false)
     @CacheEvict(value = {
@@ -515,19 +514,13 @@ public class CertifiedProductManagerImpl implements CertifiedProductManager {
 
         // qms
         if (pendingCp.getQmsStandards() != null && pendingCp.getQmsStandards().size() > 0) {
-            for (PendingCertifiedProductQmsStandardDTO qms : pendingCp.getQmsStandards()) {
+            for (PendingCertifiedProductQmsStandardDTO pendingQms : pendingCp.getQmsStandards()) {
                 CertifiedProductQmsStandardDTO qmsDto = new CertifiedProductQmsStandardDTO();
-                if (qms.getQmsStandardId() == null) {
-                    QmsStandardDTO toAdd = new QmsStandardDTO();
-                    toAdd.setName(qms.getName());
-                    toAdd = qmsDao.create(toAdd);
-                    qmsDto.setQmsStandardId(toAdd.getId());
-                } else {
-                    qmsDto.setQmsStandardId(qms.getQmsStandardId());
-                }
+                QmsStandardDTO qms = qmsDao.findOrCreate(pendingQms.getQmsStandardId(), pendingQms.getName());
+                qmsDto.setQmsStandardId(qms.getId());
                 qmsDto.setCertifiedProductId(newCertifiedProduct.getId());
-                qmsDto.setApplicableCriteria(qms.getApplicableCriteria());
-                qmsDto.setQmsModification(qms.getModification());
+                qmsDto.setApplicableCriteria(pendingQms.getApplicableCriteria());
+                qmsDto.setQmsModification(pendingQms.getModification());
                 cpQmsDao.createCertifiedProductQms(qmsDto);
             }
         }
@@ -636,18 +629,12 @@ public class CertifiedProductManagerImpl implements CertifiedProductManager {
                     }
 
                     if (certResult.getUcdProcesses() != null && certResult.getUcdProcesses().size() > 0) {
-                        for (PendingCertificationResultUcdProcessDTO ucd : certResult.getUcdProcesses()) {
+                        for (PendingCertificationResultUcdProcessDTO pendingUcd : certResult.getUcdProcesses()) {
                             CertificationResultUcdProcessDTO ucdDto = new CertificationResultUcdProcessDTO();
-                            if (ucd.getUcdProcessId() == null) {
-                                UcdProcessDTO newUcd = new UcdProcessDTO();
-                                newUcd.setName(ucd.getUcdProcessName());
-                                newUcd = ucdDao.create(newUcd);
-                                ucdDto.setUcdProcessId(newUcd.getId());
-                            } else {
-                                ucdDto.setUcdProcessId(ucd.getUcdProcessId());
-                            }
+                            UcdProcessDTO ucd = ucdDao.findOrCreate(pendingUcd.getUcdProcessId(), pendingUcd.getUcdProcessName());
+                            ucdDto.setUcdProcessId(ucd.getId());
                             ucdDto.setCertificationResultId(createdCert.getId());
-                            ucdDto.setUcdProcessDetails(ucd.getUcdProcessDetails());
+                            ucdDto.setUcdProcessDetails(pendingUcd.getUcdProcessDetails());
                             certDao.addUcdProcessMapping(ucdDto);
                         }
                     }
@@ -991,8 +978,8 @@ public class CertifiedProductManagerImpl implements CertifiedProductManager {
     }
 
     @Override
-    @PreAuthorize("hasRole('ROLE_ADMIN') or " + "( (hasRole('ROLE_ACB_STAFF') or hasRole('ROLE_ACB_ADMIN'))"
-            + "  and hasPermission(#acbId, 'gov.healthit.chpl.dto.CertificationBodyDTO', admin)" + ")")
+    @PreAuthorize("hasRole('ROLE_ADMIN') or " + "hasRole('ROLE_ACB')"
+            + "  and hasPermission(#acbId, 'gov.healthit.chpl.dto.CertificationBodyDTO', admin)")
     @Transactional(readOnly = false)
     public void sanitizeUpdatedListingData(Long acbId, CertifiedProductSearchDetails listing)
             throws EntityNotFoundException {
@@ -1036,8 +1023,8 @@ public class CertifiedProductManagerImpl implements CertifiedProductManager {
     }
 
     @Override
-    @PreAuthorize("hasRole('ROLE_ADMIN') or " + "( (hasRole('ROLE_ACB_STAFF') or hasRole('ROLE_ACB_ADMIN'))"
-            + "  and hasPermission(#acbId, 'gov.healthit.chpl.dto.CertificationBodyDTO', admin)" + ")")
+    @PreAuthorize("hasRole('ROLE_ADMIN') or " + "hasRole('ROLE_ACB')"
+            + "  and hasPermission(#acbId, 'gov.healthit.chpl.dto.CertificationBodyDTO', admin)")
     @Transactional(rollbackFor = {
             EntityRetrievalException.class, EntityCreationException.class, JsonProcessingException.class,
             AccessDeniedException.class, InvalidArgumentsException.class
@@ -1097,7 +1084,7 @@ public class CertifiedProductManagerImpl implements CertifiedProductManager {
                 }
             } else if (!Util.isUserRoleAdmin() && !Util.isUserRoleAcbAdmin()) {
                 LOGGER.error("User " + Util.getUsername()
-                        + " does not have ROLE_ADMIN or ROLE_ACB_ADMIN and cannot change the status of developer for certified product with id "
+                        + " does not have ROLE_ADMIN or ROLE_ACB and cannot change the status of developer for certified product with id "
                         + listingId);
                 throw new AccessDeniedException(
                         "User does not have admin permission to change " + cpDeveloper.getName() + " status.");